refactor: keep routing separate
